Switch to base16-flat color scheme
I've largely dropped support for consoles that don't support 256 colors
since I can't remember when I used one last. Profiles for iTerm2 and
Terminal.app are in the profiles dir. `base16-shell` is in resources and
sets terminal colors for using a 256-color terminal theme (all of the
main profiles in the profiles dir are 256-color).