Kill Ctrl-Backspace keys in tecla and inputrc
Generally I've found terminals send ^? for backspace and ^H for
Ctrl-Backspace. This allowed me to map ^H to kill-backwards-word.
Unfortunately, tmux seems to workaround the possibility that a session
could be connected by two terminals that differ in this case by
rewriting all ^? to ^H and treating that as backspace in the TERM.

This seemed to only matter in some certain situations? I noticed in
GDB's TUI that backspace was deleting words. Either way, the accepted
solution is just to get used to Ctrl-W or M-Backspace.