Force tmux truecolors, don't use tmux control seq for solar mode 3
Tmux was changing RGB values to the closest xterm 256 color when
truecolor wasn't enabled. Turns out when this is solved, Tmux handles
the normal control sequences correctly, so we don't have to pass through
the sequences or worry about running in Tmux at all.

Tested on windows terminal, putty (Win 11)